refactor(tests): collapse UnifiedGestures into Gestures
The Detox/Appium strategy split is gone; Gestures now calls AppiumGestures directly as the single page-object entry point.
