+++ tests/framework/fixtures/playwright/deviceProvider.fixture.ts
@@ -2,7 +2,6 @@ import type { Fixtures, FullProject, WorkerInfo } from '@playwright/test';
 import { createServiceProvider, type ServiceProvider } from '../../services';
 import type { WebDriverConfig } from '../../types.ts';
 import { createAppiumLogger } from '../../appiumLogger.ts';
-import UnifiedGestures from '../../UnifiedGestures.ts';
 import type { SharedAppiumSession, WorkerLevelFixtures } from './types.ts';
 
 const logger = createAppiumLogger('deviceProvider');
@@ -67,7 +66,6 @@ export const workerDeviceProviderFixture: Fixtures<
 
       try {
         delete globalThis.driver;
-        UnifiedGestures.resetStrategy();
       } catch (error) {
         logger.error(
           'Failed to clear global driver on worker teardown:',
