+++ tests/framework/fixtures/playwright/driver.fixture.ts
@@ -15,7 +15,6 @@ import {
   requestSharedSessionRecreate,
 } from '../../services/appium/sessionRecovery.ts';
 import { createAppiumLogger } from '../../appiumLogger.ts';
-import UnifiedGestures from '../../UnifiedGestures.ts';
 import type { ServiceProvider } from '../../services';
 import { isAppiumSessionReuseEnabled } from './sessionReuse.ts';
 
@@ -114,7 +113,6 @@ export const driverFixture = {
       await configureImplicitWait(drv, implicitMs);
 
       globalThis.driver = drv;
-      UnifiedGestures.resetStrategy();
 
       const platformName = (await drv.capabilities)?.platformName;
       const windowSize = await drv.getWindowSize();
@@ -240,7 +238,6 @@ export const driverFixture = {
 
         try {
           delete globalThis.driver;
-          UnifiedGestures.resetStrategy();
         } catch (error) {
           logger.error('Failed to clean up global driver:', error);
         }
