+++ tests/tools/e2e-ai-analyzer/modes/select-tags/test-infrastructure-paths.ts
@@ -15,8 +15,6 @@ const FRAMEWORK_INFRA_EXACT_FILES = new Set([
   'tests/framework/Assertions.ts',
   'tests/framework/EncapsulatedElement.ts',
   'tests/framework/Gestures.ts',
-  'tests/framework/GestureStrategy.ts',
-  'tests/framework/UnifiedGestures.ts',
   'tests/framework/Matchers.ts',
   'tests/framework/Utilities.ts',
   'tests/framework/index.ts',
